klimaanlage mit den Funktionen getTemperature und setTemperature implementiert
