+++ src/Erde/helpFunctions/helpFunctions.js
