+++ src/Erde/helpFunctions/requestFunction.js
