+++ src/Erde/helpFunctions/responseToMars.js
